fix fonts

+++ init/02_ubuntu_apt.sh
@@ -12,26 +12,38 @@ sudo apt-get install -y \
   tree \
   oathtool \
   vim \
-  python3-pip
+  python3-pip \
+  powerline
 
 # Create installers foler
 mkdir -p "$DOTFILES/caches/installers/"
 
-# Remove LibreOffice
-sudo apt-get remove --purge libreoffice* -y
-sudo apt-get clean -y
-sudo apt-get autoremove -y
+if is_ubuntu_desktop; then
+  # Remove LibreOffice
+  sudo apt-get remove --purge libreoffice* -y
+  sudo apt-get clean -y
+  sudo apt-get autoremove -y
 
-# Visual Studio Code
-# https://code.visualstudio.com/docs/setup/linux
-wget -O "$DOTFILES/caches/installers/vscode.deb" https://vscode-update.azurewebsites.net/latest/linux-deb-x64/stable
-sudo dpkg -i "$DOTFILES/caches/installers/vscode.deb"
-sudo apt-get install -fy
+  # Visual Studio Code
+  # https://code.visualstudio.com/docs/setup/linux
+  wget -O "$DOTFILES/caches/installers/vscode.deb" https://vscode-update.azurewebsites.net/latest/linux-deb-x64/stable
+  sudo dpkg -i "$DOTFILES/caches/installers/vscode.deb"
+  sudo apt-get install -fy
 
-# Chrome
-wget -O "$DOTFILES/caches/installers/chrome.deb" https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
-sudo dpkg -i "$DOTFILES/caches/installers/chrome.deb"
-sudo apt-get install -fy
+  # Chrome
+  wget -O "$DOTFILES/caches/installers/chrome.deb" https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
+  sudo dpkg -i "$DOTFILES/caches/installers/chrome.deb"
+  sudo apt-get install -fy
+
+  # Slack
+  wget -O "$DOTFILES/caches/installers/slack.deb" https://downloads.slack-edge.com/linux_releases/slack-desktop-2.9.0-amd64.deb
+  sudo dpkg -i "$DOTFILES/caches/installers/slack.deb"
+  sudo apt-get install -fy
+
+  # Gnome extensions
+  # Tiling window manager
+  git clone https://github.com/vibou/vibou.gTile.git ~/.local/share/gnome-shell/extensions/gTile@vibou
+fi
 
 # yarn
 # https://yarnpkg.com/en/docs/install#linux-tab
@@ -41,7 +53,7 @@ sudo apt-get update -y && sudo apt-get install yarn -y
 
 # docker
 # https://docs.docker.com/engine/installation/linux/docker-ce/ubuntu/#install-docker-ce
-sudo apt-get install \
+sudo apt-get install -y \
   apt-transport-https \
   ca-certificates \
   curl \
@@ -51,7 +63,7 @@ sudo add-apt-repository \
   "deb [arch=amd64] https://download.docker.com/linux/ubuntu \
   $(lsb_release -cs) \
   stable"
-sudo apt-get update
+sudo apt-get update -y
 sudo apt-get install docker-ce -y
 # Add current user to docker group
 sudo groupadd docker
@@ -61,11 +73,6 @@ newgrp docker
 # docker-compose
 sudo curl -L https://github.com/docker/compose/releases/download/latest/docker-compose-`uname -s`-`uname -m` -o /usr/local/bin/docker-compose
 sudo chmod +x /usr/local/bin/docker-compose
-
-# Slack
-wget -O "$DOTFILES/caches/installers/slack.deb" https://downloads.slack-edge.com/linux_releases/slack-desktop-2.9.0-amd64.deb
-sudo dpkg -i "$DOTFILES/caches/installers/slack.deb"
-sudo apt-get install -fy
 
 # pass and pass otp
 # https://github.com/tadfisher/pass-otp
@@ -79,7 +86,3 @@ git clone https://github.com/tadfisher/pass-otp "$DOTFILES/caches/installers/pas
 builtin cd "$DOTFILES/caches/installers/pass-otp"
 sudo make install
 builtin cd -
-
-# Gnome extensions
-# Tiling window manager
-git clone https://github.com/vibou/vibou.gTile.git ~/.local/share/gnome-shell/extensions/gTile@vibou
